Add files via upload
It is required to maintain a library database using a number of lists
as mentioned:
The list BOOKS contains the information like title, accession
number, and tag field (to indicate whether a book is issued or not)
for all the books in a library. Note that each book can be there in
multiple copies, but there accession numbers are different.
Another list SUBSCRIBERS will contain the name, borrower
number and the list of books (with date of issues) he has issued.
Assume that a subscriber can issue up to five books at the most and
no two copies of the same book.
(a) Design a suitable data structure using single linked list.
(b) Write a menu driven program using C for the following:
(i) To issue a book
(ii) Return a book
(iii) Show the list of books issued by a subscriber
(iv) Given a title, find out to whom it has issued.